requirements.txt
emoji
pyfiglet
inflect
requests
tabulate
Pillow

